requirements.txt
pytest
torch
torchvision
numpy
Pillow
scikit-image 
scikit-learn
pandas
opencv-python
tensorboard
pyyaml
matplotlib
tqdm
torchsampler